Converts role build-generic-pkg -> build-debian-package
Converts OSSEC deb creation to new generalized role. We're now using the
same role to build the following Debian packages:

  * securedrop-app-code
  * securedrop-ossec-agent
  * securedrop-ossec-server

The base dependencies for OSSEC, ossec-agent and ossec-server, are still
built in an external repository, using a different role.

New role includes a debugging task to display the package name being built
during the output of the Ansible run. Wrapper roles will also clarify
the current package being built.

Already synchronizing directories of static files, but doesn't hurt to
add support for custom filepaths to `copy` in. Requires "src" and "dest"
attributes for the filepaths provided in a list. This is mostly
necessary due to the technical debt of storing the DEBIAN control files
for the built packages in a higher-level directory in the securedrop git
repository.

Also carefully manipulates the rsync command used by `synchronize` to
ensure correct permissions and ownership on the dest files.