Updates template to iterate over control field dict
Also creates list of listvars for comma-separation. Certain Debian
control fields can take multiple arguments, and should be
comma-separated. Currently implementing by forcing expectation of a list
and comma-separating the list inside the template.

If there's a way to type-check vars in Ansible/Jinja, so dict values can
be inspected and handled according to list/string, I don't know it.